--- server/src/common/dto/vote.go
@@ -1,3 +1,0 @@
-package dto
-
-// Vote is the response for all vote requests.
